+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Fixed an issue where texture layers would only assign the first channel.
 * Fixed an issue where not all meshes would render in the color debug shading mode.
+* Fixed an issue where `xc3_shader::graph::Graph` did not properly handle GLSL parameter syntax.
+* Fixed an issue where the `glsl-dependencies` command for xc3_shader did not work for variables without channels.
 
 ## 0.13.0 - 2024-09-18
 ### Added